Make BigFraction.doubleValue() stable for large fractions
If num or denom are very large they are converted into +/- Infinite by
BigInteger.doubleValue(). The subsequent division results in a non-finite
result.

If both are very large yet within a "small" factor of each other, the result
should be representable. However, because they are converted to non-finite
values before the division, the result comes back as non-finite.

We can avoid this by performing the division in BigDecimal prior to converting
to double.