Bug fix: Handle 'unknown' SMT results during invariant generation
This fixes a bug reported by Anthony Aiello and Aaron Fifarek. When
the invariant generator received an 'unknown' result it was treating
it like an 'unsat' result. This caused properties to be erroneously
reported as true.

This bug only appeared in the inductive step part of the invariant
generation. Handling of 'unknown' in the base step was fixed in
99394a840a8b47ef8153fd224548c4b9550cd820 to address issue #20. This
looks like a case that slipped through at that time.